% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjMat}
\alias{adjMat}
\title{Compute adjacency matrix of a decomposable graph}
\usage{
adjMat(graph, p)
}
\arguments{
\item{graph}{Decomposable graph as in \code{data(graphs5)}.}

\item{p}{of vertices of the graph.}
}
\value{
Adjacency matrix.
}
\description{
Compute adjacency matrix of a decomposable graph
}
