% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmcr.R
\name{diagnostics}
\alias{diagnostics}
\title{Compute MCMC convergence diagnostics for MCMC trace}
\usage{
diagnostics(traces)
}
\arguments{
\item{traces}{output from the \code{MCMCtrace()} function.}
}
\description{
Computes MCMC convergence diagnostics, R hat, effective sample size, and number of iterations, for MCMC trace obtained from the \code{MCMCtrace()} function.
}
