% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestDecisionModels.R
\name{HarvDec2}
\alias{HarvDec2}
\title{Harvest decision model - when measures of hunters frusration is available}
\usage{
HarvDec2(TAC_t0 = 2, qu = 0.2, UR_targ = 0.8, UR_est = 0.7)
}
\arguments{
\item{TAC_t0}{The total allowable catch (TAC) the previous year}

\item{qu}{Scaling paramter deciding how much the managers will adjus qouta as a response to hunters frustration,}

\item{UR_targ}{The target level of frustrain that the managers are aimming for}

\item{UR_est}{The measured or estimated level of frustraion among hunters}
}
\description{
Harvest dection (qouta or effort) is decided based on information about
the level of frustration among hunters.
}

