% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestDecisionModels.R
\name{HarvDec_E}
\alias{HarvDec_E}
\title{Harvest decision model - when measures of hunters frustration is available}
\usage{
HarvDec_E(TAC_t0 = 2, qu = 0.2, UR_targ = 0.8, UR_est = 0.7)
}
\arguments{
\item{TAC_t0}{The total allowable catch (TAC) the previous year}

\item{qu}{Scaling parameter deciding how much the managers will adjust qouta 
as a response to hunters frustration,}

\item{UR_targ}{The target level of frustration that the managers are aiming for}

\item{UR_est}{The measured or estimated level of frustraion among hunters}
}
\value{
TAC Total allowable catch emerging from model and parameters
}
\description{
Harvest size (qouta or effort) is based on information about
the level of frustration among hunters.
THIS IS WORK IN PROGRESS - JUST A VERY ROUGH IDEA!
}
