% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImplementationModels.R
\name{Impl2}
\alias{Impl2}
\title{Schaefer's CPUE model}
\usage{
Impl2(Effort = 3, Pop_state = 100, q = 0.1)
}
\arguments{
\item{Effort}{defines the effort, that will be multiplied by q to estimate total catch}

\item{Pop_state}{Population abundance (or biomass)}

\item{q}{Catchability coeficient}
}
\value{
H_I Number of harvested individuals (or harvested biomass)
}
\description{
Schaefer`s catch pr unit effort (CPUE) model. Assumes that catchability (q) is constant
and not dependent on population density.
}
