% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImplementationModels.R
\name{Impl_CM}
\alias{Impl_CM}
\title{Csirke_MacCall`s CPUE model}
\usage{
Impl_CM(Effort = 3, alpha = 0.13, beta = 0.1, Pop_state = 10)
}
\arguments{
\item{Effort}{defines the effort, that will be multiplied by q to estimate total catch}

\item{alpha}{Parameter is catchability model}

\item{beta}{Parameter in catchability model}

\item{Pop_state}{Population abundance (or biomass)}
}
\value{
H_I Number of harvested individuals (or harvested biomass)
}
\description{
Csirke_MacCall`s catch pr unit effort (CPUE) model. Assumes that catchability (q) is dependent on
abundance (or population density), most often increasing as abundance decrease.
}
