% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImplementationModels.R
\name{Impl_rd}
\alias{Impl_rd}
\title{Implementation of harvest quota}
\usage{
Impl_rd(TAC = 10, ModType = "A", p = 0.7)
}
\arguments{
\item{TAC}{Total allowable catch. Usually the output from a harvest decision model. Note that TAC must be an integer number
when ModType="A".}

\item{ModType}{Type of probability model, being either A) Binomial or B) Poisson}

\item{p}{Probability of harvest in the binomial model, or "bias" in the Poisson model}
}
\value{
H_I Number of harvested animals
}
\description{
The simplest form of an implementation model - basically just a random draw from a
binomial (ModType="A") or a Poisson (ModType="B") distribution.
}
