% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImplementationModels.R
\name{Impl_schaefer}
\alias{Impl_schaefer}
\title{Schaefer's CPUE model}
\usage{
Impl_schaefer(Effort = 3, Pop_state = 100, q = 0.1)
}
\arguments{
\item{Effort}{Define the effort, that together with the catchability constant (q) and 
abundance (or population density) will determine total catch}

\item{Pop_state}{Population abundance (or biomass)}

\item{q}{Catchability coeficient}
}
\value{
H_I Number of harvested individuals (or harvested biomass)
}
\description{
Schaefer`s catch pr unit effort (CPUE) model. Assumes that catchability (q) is constant
and not dependent on abundance (or density).
}
