% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImplementationModels.R
\name{Impl_SM2}
\alias{Impl_SM2}
\title{Csirke_MacCall`s CPUE model}
\usage{
Impl_SM2(Effort = 3, alpha = 0.13, beta = 0.1, Pop_state = 10,
  Est_Pop_state = 10, c = 10)
}
\arguments{
\item{Effort}{defines the effort, that will be multiplied by q to estimate total catch}

\item{alpha}{Parameter is catchability model}

\item{beta}{Parameter in catchability model}

\item{Pop_state}{Population abundance (or biomass)}

\item{Est_Pop_state}{Estimated population size emerging from observation model (i.e. monitoring)}

\item{c}{Threshold where harvesting is abondend}
}
\value{
H_I Number of harvested individuals (or harvested biomass)
}
\description{
Csirke_MacCall`s catch pr unit effort (CPUE) model. Assumes that catchability (q) increase as
population density (or biomass) decreases.
}
\details{
Modified to include lower threshold where harvest
is abondened, determined from an observation of population state erging from an observation
model. This is a shortcut, because such thresholds are most often
determined in a harvest decision model.
}
