% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{independence}
\alias{independence}
\title{Fit an independence model}
\usage{
independence(data)
}
\arguments{
\item{data}{object of class "MSEdata" representing list inclusion pattern counts.}
}
\value{
Object of class "independence" containing the fitted model "fit" and the arguments "args" passed to this function.
}
\description{
Fit a (sparse) independence model to the data using the \code{SparseMSE::modelfit()} function.
}
\examples{
indep_fit <- independence(UK)
estimates(indep_fit)
}
