% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmcr.R
\name{lcmcr}
\alias{lcmcr}
\title{Fit a latent class model}
\usage{
lcmcr(data, K=10, seeds=1:200, lazy=TRUE, ...)
}
\arguments{
\item{data}{object of class "MSEdata" representing list inclusion pattern counts.}

\item{K}{maximum number of latent classes.}

\item{seeds}{numeric vector of RNG seeds for the MCMC chains. A chain is initialized for vector element.}

\item{lazy}{whether or not to wait before initializing the chains with lcmcr objects. Default is TRUE.}

\item{...}{other parameters passed to the \code{LCMCR::lcmCR()} function for each chain.}
}
\description{
Fit the latent class model of Manrique-Vallier (2016). This function initializes the MCMC chains and sampling is postponed until a call to the \code{estimates()} function. By default, 200 independent chains are initialized.
}
\examples{
lcmcr_fit <- lcmcr(UK)
estimates(lcmcr_fit)
}
