% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{listnames<-}
\alias{listnames<-}
\title{Set list names}
\usage{
`listnames<-`(data, value)
}
\arguments{
\item{data}{Data of type \code{MSEdata}.}

\item{value}{character vector of list names of length \code{nlists()}}
}
\description{
Set list names
}
