% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{make_depth_poly}
\alias{make_depth_poly}
\title{EXPERIMENTAL}
\usage{
make_depth_poly(bath, depth_range = c(0, -10), domain_outer)
}
\description{
\code{make_depth_poly} create depth polygon by joining depth contour lines
}
