% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmcr.R
\name{MCMCtrace}
\alias{MCMCtrace}
\title{Get MCMC traces for LCMCR}
\usage{
MCMCtraces(fit, burnin = 0, nSamples = 100, thinning = 1000, mc.cores=detectCores())
}
\arguments{
\item{fit}{object of class "lcmcr".}

\item{burnin}{number of burn-in iterations.}

\item{nSamples}{number of samples to be returned.}

\item{thinning}{number of iterations between each sample.}

\item{mc.cores}{number of cores to use for computation.}
}
\description{
Run MCMC and return traces for the "n0", "prob_zero" and "k_star" parameters of LCMCR objects. For each of these parameters, a matrix is returned where each row represents an iteration and each column a chain.
}
