% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{mergeLists}
\alias{mergeLists}
\title{Merge lists}
\usage{
mergeLists(data, ...)
}
\arguments{
\item{data}{Data of type \code{MSEdata}.}

\item{...}{Character vectors representing sets of lists to be merged.}
}
\description{
Merge lists
}
