% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{MSEdata}
\alias{MSEdata}
\title{MSE data format}
\usage{
MSEdata(data, colname = "count")
}
\arguments{
\item{data}{Original MSE dataframe. It should contain a column named representing
the observed counts of capture patterns, as well as columns representing
the different lists, as follows:}

\item{colname}{Name of the column representing the observed counts. Default is "count".
\preformatted{         c1    c2 count
         0     1     7
         1     0     3
         1     1     4}}
}
\description{
The function \code{MSEdata()} transforms an existing dataframe to the "MSE" format,
ensuring it contains a "count" column and that the other columns refer to
inclusion (1) or exclusion (0) on a set of lists.
}
\details{
Zero counts of unobserved capture patterns are added and duplicates capture patterns
are aggregated.
}
