% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSETools.R
\docType{package}
\name{MSETools}
\alias{MSETools}
\alias{MSETools-package}
\title{\pkg{MSETools}}
\description{
MSETools provides a unified interface to multiple systems estimation (MSE) software. It implements best usage practices and computational speedups. Data from multiple system estimation studies of human trafficking has been reproduced for illustrations and analyses.
}
\author{
\strong{Maintainer}: Olivier Binette \email{olivier.binette@gmail.com} (\href{https://orcid.org/0000-0001-6009-5206}{ORCID})

}
