% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationModels.R
\name{obs_mod1}
\alias{obs_mod1}
\title{Simple(st) observation model}
\usage{
obs_mod1(scale = "Abund", value = 1000, bias = 1, cv = 0.2,
  LogNorm = "ND")
}
\arguments{
\item{scale}{Indicates if you are modelling density ("Dens") or Abundance ("Abund").}

\item{value}{The true value of abundance/density}

\item{bias}{Relative bias, where 1=no bias, 0-0.99 = bias low, >1 = bias high}

\item{cv}{Coef of variation of the monitoring program/survey - in relation to true value.}

\item{LogNorm}{Indicates if a lognormal ("LND") or normal ("ND") density model is used}
}
\value{
obs1 Mean value of the observation (i.e. observation)
}
\description{
A simplistic observation model, based on the lognormal or normal distribution.
}
