% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{om}
\alias{om}
\title{Run operating model}
\usage{
om(data_om, a = NULL, seed = NULL, verbose = TRUE)
}
\arguments{
\item{data_om}{tagged list containing model parameters}

\item{a}{\code{NULL} or the number of time periods to run OM before updating management plan}

\item{seed}{seed for random number generation}

\item{verbose}{if \code{TRUE} then runtime feedback is provided}
}
\value{
object of class \code{om}, containing the om output:

\describe{
   \itemize{
     \item \code{data_om}: updated list of model parameters
     \itemize{
       \item ...
     }
     \item \code{om_rep}: reported values from \code{C++} file
     \itemize{
       \item \code{SSB0_c}: unfished spawning-stock-biomass for each species
       \item \code{N_csl}: numbers of individuals-at-size from last simulated time period. This variable is used for model re-entry when \code{a != NULL}
       \item \code{B_cst}: biomass by species, time and spatial location
       \item \code{ncatch_cstl}: number of harvested individuals-at-size attributed to each mesh loation
       \item \code{c_i}: zero-indexed species associated with the ith harvest tow
       \item \code{f_i}: zero-indexed fishing grid cell associated with the ith harvest tow
       \item \code{t_i}: zero-indexed time period associated with the ith harvest tow
       \item \code{catch_i}: harvested biomass from ith harvest tow
       \item \code{ncatch_il}: harvested numbers-at-size from ith harvest tow
     }
     \item \code{harvest_time}: information on model execution time
     \itemize{
       \item ...
     }
   }
}
}
\description{
\code{om} runs the \code{operating model}
}
