% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEdata.R
\name{omit}
\alias{omit}
\title{Omit lists from MSE data}
\usage{
omit(data, lists)
}
\arguments{
\item{data}{Data of type \code{MSEdata}.}

\item{lists}{Character vector of lists to be omitted.}
}
\description{
Omit lists from MSE data
}
