% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModels.R
\name{PopMod_Gomp}
\alias{PopMod_Gomp}
\title{Gompertz population model - harvest before density dependent growth}
\usage{
PopMod_Gomp(X_t0 = 100, sigma2_e = 0.2, N_Harv = 20, K = 200,
  r_max = 1)
}
\arguments{
\item{X_t0}{Population size (X) at time t=0 (i.e. at the time step before the function applies)}

\item{sigma2_e}{environmental stochasticity}

\item{N_Harv}{Number of individuals harvested (in year t0)}

\item{K}{Carrying capacity of the theta-logistic model}

\item{r_max}{Maximum growth rate (i.e. intrinsic growth rate)}
}
\value{
Data frame containing the following columns:

eps: Value of the stochastic component in year t

X_star Post-harvest population size

X_t1 Population size (X) in year t=1

lam Population growth rate (lambda) in year t
}
\description{
A gompertz-type population model, including the possibility for harvest. Harvest is applied
as a pulse removing a pre-determined number of individuals. After harvest the (model) population
grows according to the parameters in the model.
}
