% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModels.R
\name{PopMod_TLpost}
\alias{PopMod_TLpost}
\title{Theta-logistic population model - harvest after density dependent growth}
\usage{
PopMod_TLpost(X_t0 = 100, sigma2_e = 0.2, N_Harv = 20, theta = 1,
  K = 200, r_max = 1.3)
}
\arguments{
\item{X_t0}{Population size (X) at time t=0 (i.e. at the time step before the function applies)}

\item{sigma2_e}{environmental stoachasticity}

\item{N_Harv}{Number of individuals harvested (in year t0)}

\item{K}{Carrying capacity of the theta-logistic model}

\item{r_max}{Maximum growth rate (i.e. intrinsic growth rate)}
}
\value{
Data frame containing the following columns:

eps: Value of the stochastic component in year t

X_star Post-harvest population size

r Population growth rate

X_t1 Population size (X) in year t=1
}
\description{
A theta-logistic population model, including the possibility for harvest. Harvest is applied
as a pulse, removing a pre-determined number of individuals. Before harvest is applied, the population
grows according to the parameters governing the dynamics of the (model) population. When theta = 1, this become the 
conventional Ricker model.
}
