% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModels.R
\name{PopMod_TLpre}
\alias{PopMod_TLpre}
\title{Theta-logistic population model - harvest before density dependent growth}
\usage{
PopMod_TLpre(X_t0 = 100, sigma2_e = 0.2, N_Harv = 20, K = 200,
  theta = 1, r_max = 1)
}
\arguments{
\item{X_t0}{Population size (X) at time t=0 (i.e. at the time step before the function applies)}

\item{sigma2_e}{environmental stoachasticity}

\item{N_Harv}{Number of individuals harvested (in year t0)}

\item{K}{Carrying capacity of the theta-logistic model}

\item{r_max}{Maximum growth rate (i.e. intrinsic growth rate)}
}
\value{
Data frame containing the following columns:

eps: Value of the stochastic component in year t

X_star Post-harvest population size

r Population growth rate

X_t1 Population size (X) in year t=1
}
\description{
A theta-logistic (theta-Ricker) population model, including the possibility for harvest. Harvest is applied
as a pulse, removing a pre-determined number of individuals. After harvest the population
grows according to the parameters in the model. Environmental stochasticity (process error) is modelled as an 
 additive effect on r(t). When theta = 1, the model it the conventional Ricker model .
}
