% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationModels.R
\name{PopMod_TLvar}
\alias{PopMod_TLvar}
\title{Theta-logistic population model - variant}
\usage{
PopMod_TLvar(X_t0 = 100, sigma2_e = 0.2, N_Harv = 20, K = 200,
  theta = 1, r_max = 1)
}
\arguments{
\item{X_t0}{Population size (X) at time t=0 (i.e. at the time step before the function applies)}

\item{sigma2_e}{environmental stoachasticity}

\item{N_Harv}{Number of individuals harvested (in year t0)}

\item{K}{Carrying capacity of the theta-logistic model}

\item{r_max}{Maximum growth rate (i.e. intrinsic growth rate)}
}
\value{
Data frame containing the following columns:

eps: Value of the stochastic component in year t

X_star Post-harvest population size

r Population growth rate

X_t1 Population size (X) in year t=1
}
\description{
A theta-logistic population model Variant (See Aanes et al. 2002 for model formulation), including the possibility for harvest. Harvest is 
applied as a pulse removing a pre-determined number of individuals. After harvest the population
grows according to the parameters governing the dynamics of the (model) population. Note that this type of model become the Gompertz model
as theta approaches 0.
}
