% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{print.om}
\alias{print.om}
\title{Print data}
\usage{
\method{print}{om}(x, units = "kg", ...)
}
\arguments{
\item{x}{Output from \code{\link{om}}}

\item{...}{Not used}
}
\description{
Print output from \code{\link{om}}
}
