% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestDecisionModels.R
\name{HarvDec1}
\alias{HarvDec1}
\title{General harvest decision model}
\usage{
HarvDec1(HD_type = "A", c = 1000, qu = 0.2, PopState_est = 100)
}
\arguments{
\item{HD_type}{Type of harvest decision model, being one of three:
A: Proportional harvest (i.e. constant harvest mortality)
B: Threshold harvest (no harvest when bellow threshold c)
C: constant quota harvest (i.e. remove same number of individuals regardless of pop. size)}

\item{c}{Threshold in the threshold harvest model}

\item{qu}{Depending on type of harvest model: for proportional harvest, this is the harvest rate,
for constant quota this is the number of individuals removed. Not relevant for (absolute) threshold
harvesting}

\item{PopState_est}{The (estimated) population size available to managers when making decision}
}
\description{
Select harvest decision model (proportional harvest, constant quota or threshold harvesting)
to be applied. Each model are supplied with a set of control parameters.
}

